Feat: Visualize insertion operation in Trie

+++ src/Trie/TrieAnimStep.cpp
@@ -8,12 +8,12 @@ TrieAnimStep::TrieAnimStep(TrieAnimType type, std::string description, std::vect
 	: type(type), description(description), highlightCodeLineIndex(highlightCodeLineIndex)
 {}
 
-TrieAnimStep::TrieAnimStep(TrieAnimType type, std::string description, std::vector<int> highlightCodeLineIndex, int curID)
+TrieAnimStep::TrieAnimStep(TrieAnimType type, std::string description, std::vector<int> highlightCodeLineIndex, uint64_t curID)
 	: type(type), description(description), highlightCodeLineIndex(highlightCodeLineIndex)
 	, hasCurID(true), curID(curID)
 {}
 
-TrieAnimStep::TrieAnimStep(TrieAnimType type, std::string description, std::vector<int> highlightCodeLineIndex, int curID, int oldTreeSnapshotIndex)
+TrieAnimStep::TrieAnimStep(TrieAnimType type, std::string description, std::vector<int> highlightCodeLineIndex, uint64_t curID, int oldTreeSnapshotIndex)
 	: type(type), description(description), highlightCodeLineIndex(highlightCodeLineIndex)
 	, hasCurID(true), curID(curID), oldTreeSnapshotIndex(oldTreeSnapshotIndex)
 {}
